Adding some DOM interface implementations
git-svn-id: https://svn.lojjic.net/xul/trunk@536 8f8d9d7b-adf4-0310-9a85-ad1d9f542d00

+++ xul-dom-api/src/main/java/net/lojjic/xul/XULSelectControlElement.java
@@ -5,16 +5,16 @@ public interface XULSelectControlElement extends XULControlElement {
 	XULSelectControlItemElement getSelectedItem();
 	void setSelectedItem(XULSelectControlItemElement selectedItem);
 	
-	long getSelectedIndex();
-	void setSelectedIndex(long selectedIndex);
+	int getSelectedIndex();
+	void setSelectedIndex(int selectedIndex);
 	
 	String getValue();
 	void setValue(String value);
 	
 	XULSelectControlItemElement appendItem(String label, String value);
 	
-	XULSelectControlItemElement insertItemAt(long index, String label, String value);
+	XULSelectControlItemElement insertItemAt(int index, String label, String value);
 	
-	XULSelectControlItemElement removeItemAt(long index);
+	XULSelectControlItemElement removeItemAt(int index);
 	
 }
